Fix PR2510
dbtables.get_xml() does not respect connection argument.  this fix adds the
ability to pass a connection argument to the DBTable class at instance
creation time to override the class attribute, and get_xml() uses this to
fix its problem.  adding the connection keyword argument might not work in
the long run, but it takes the code in the direction of the ideal goal of
eliminating the "global" connection class attribute altogether.