glue.ligolw.dbtables:  checksum temporary files
This patch adds checksum comparison to the "move file from scratch space
back to permanent location" code to ensure the checksum in the permanent
location matches the checksum on the scratch disk.  This triples (in the
best case) the I/O involved in replacing the file to its permanent location
so this is not intended to be a permanent feature of the code, it is only
here temporarily while the file corruption on the LDG clusters is
diagnosed.