glue.ligolw:  allow pluggable namespace handling
this patch makes further progress on the problem of module imports having
side effects.  this removes the endElement() handlers from the content
handler class to the respective element classes themselves.  the
startElement() handlers are now explicitly told the parent element so none
of them need to access the guts of the content handler object (which, in
principle, allows them to be removed from the content handler as well, but
that is not done at this time).

this clean-up makes plugging in support for other namespaces possible.