Add ability for ligolw_segments_from_cats to only run of specific ifos or specific veto categories.
This allows us to only run certain combinations if needed, and stops the code from only recognizing values *in* the veto-definer file. For e.g. for ER4 we have to add "FAKE" veto categories at 2,3,4 and 5 to get the code to print blank veto xml files for these categories. Ihope barfs if these files are missing.